+++ src/eval/requirement/extra/data.h
@@ -32,7 +32,7 @@ public:
     ExtraData(const std::string& name, 
               const std::string& short_name, 
               const std::string& group_name,
-              float prob, 
+              double prob, 
               COMPARISON_TYPE prob_check_type, 
               EvaluationManager& eval_man,
               float min_duration,
