+++ src/eval/requirement/extra/track.cpp
@@ -39,7 +39,7 @@ namespace EvaluationRequirement
 ExtraTrack::ExtraTrack(const std::string& name, 
                        const std::string& short_name, 
                        const std::string& group_name,
-                       float prob, 
+                       double prob, 
                        COMPARISON_TYPE prob_check_type, 
                        EvaluationManager& eval_man,
                        float min_duration,
