+++ src/eval/requirement/generic/generic.h
@@ -30,7 +30,7 @@ class GenericBase : public ProbabilityBase
     GenericBase(const std::string& name, 
                 const std::string& short_name, 
                 const std::string& group_name,
-                float prob, 
+                double prob, 
                 COMPARISON_TYPE prob_check_type, 
                 EvaluationManager& eval_man);
 
@@ -53,7 +53,7 @@ public:
     GenericInteger(const std::string& name, 
                    const std::string& short_name, 
                    const std::string& group_name,
-                   float prob, 
+                   double prob, 
                    COMPARISON_TYPE prob_check_type, 
                    EvaluationManager& eval_man);
 
@@ -76,7 +76,7 @@ public:
     GenericDouble(const std::string& name, 
                   const std::string& short_name, 
                   const std::string& group_name,
-                  float prob, 
+                  double prob, 
                   COMPARISON_TYPE prob_check_type, 
                   double threshold, 
                   EvaluationManager& eval_man);
