+++ src/eval/requirement/mode_c/rocdcorrect.cpp
@@ -21,7 +21,7 @@ namespace EvaluationRequirement
 {
 
 ROCDCorrect::ROCDCorrect(const std::string& name, const std::string& short_name, const std::string& group_name,
-                         float prob, COMPARISON_TYPE prob_check_type, double threshold, EvaluationManager& eval_man)
+                         double prob, COMPARISON_TYPE prob_check_type, double threshold, EvaluationManager& eval_man)
 : GenericDouble(name, short_name, group_name, prob, prob_check_type, threshold, eval_man)
 {
     result_type_ = "ROCDCorrect";
