+++ src/eval/requirement/mom/momcorrect.cpp
@@ -22,7 +22,7 @@ namespace EvaluationRequirement
 
 
 MomLongAccCorrect::MomLongAccCorrect(const std::string& name, const std::string& short_name, const std::string& group_name,
-                                     float prob, COMPARISON_TYPE prob_check_type, EvaluationManager& eval_man)
+                                     double prob, COMPARISON_TYPE prob_check_type, EvaluationManager& eval_man)
 : GenericInteger(name, short_name, group_name, prob, prob_check_type, eval_man)
 {
     result_type_ = "MomLongAccCorrect";
@@ -39,7 +39,7 @@ MomLongAccCorrect::MomLongAccCorrect(const std::string& name, const std::string&
 
 
 MomTransAccCorrect::MomTransAccCorrect(const std::string& name, const std::string& short_name, const std::string& group_name,
-                                     float prob, COMPARISON_TYPE prob_check_type, EvaluationManager& eval_man)
+                                     double prob, COMPARISON_TYPE prob_check_type, EvaluationManager& eval_man)
     : GenericInteger(name, short_name, group_name, prob, prob_check_type, eval_man)
 {
     result_type_ = "MomTransAccCorrect";
@@ -55,7 +55,7 @@ MomTransAccCorrect::MomTransAccCorrect(const std::string& name, const std::strin
 }
 
 MomVertRateCorrect::MomVertRateCorrect(const std::string& name, const std::string& short_name, const std::string& group_name,
-                                       float prob, COMPARISON_TYPE prob_check_type, EvaluationManager& eval_man)
+                                       double prob, COMPARISON_TYPE prob_check_type, EvaluationManager& eval_man)
     : GenericInteger(name, short_name, group_name, prob, prob_check_type, eval_man)
 {
     result_type_ = "MomVertRateCorrect";
