+++ src/eval/requirement/mom/momcorrect.h
@@ -28,7 +28,7 @@ class MomLongAccCorrect : public GenericInteger
 {
   public:
     MomLongAccCorrect(const std::string& name, const std::string& short_name, const std::string& group_name,
-                      float prob, COMPARISON_TYPE prob_check_type, EvaluationManager& eval_man);
+                      double prob, COMPARISON_TYPE prob_check_type, EvaluationManager& eval_man);
 
     std::string probabilityNameShort() const override final { return "PCLAcc"; }
     std::string probabilityName() const override final { return "Probability of Correct MoM Longitudinal Acceleration"; }
@@ -40,7 +40,7 @@ class MomTransAccCorrect : public GenericInteger
 {
   public:
     MomTransAccCorrect(const std::string& name, const std::string& short_name, const std::string& group_name,
-                      float prob, COMPARISON_TYPE prob_check_type, EvaluationManager& eval_man);
+                      double prob, COMPARISON_TYPE prob_check_type, EvaluationManager& eval_man);
 
     std::string probabilityNameShort() const override final { return "PCTAcc"; }
     std::string probabilityName() const override final { return "Probability of Correct MoM Transversal Acceleration"; }
@@ -52,7 +52,7 @@ class MomVertRateCorrect : public GenericInteger
 {
   public:
     MomVertRateCorrect(const std::string& name, const std::string& short_name, const std::string& group_name,
-                       float prob, COMPARISON_TYPE prob_check_type, EvaluationManager& eval_man);
+                       double prob, COMPARISON_TYPE prob_check_type, EvaluationManager& eval_man);
 
     std::string probabilityNameShort() const override final { return "PCVRt"; }
     std::string probabilityName() const override final { return "Probability of Correct MoM Vertical Rate"; }
