+++ src/eval/requirement/position/distance.h
@@ -29,7 +29,7 @@ class PositionDistance : public ProbabilityBase
 public:
     PositionDistance(
             const std::string& name, const std::string& short_name, const std::string& group_name,
-            float prob, COMPARISON_TYPE prob_check_type, EvaluationManager& eval_man,
+            double prob, COMPARISON_TYPE prob_check_type, EvaluationManager& eval_man,
             float threshold_value, COMPARISON_TYPE threshold_value_check_type,
             bool failed_values_of_interest);
