+++ src/eval/requirement/position/latency.h
@@ -29,7 +29,7 @@ class PositionLatency : public ProbabilityBase
 public:
     PositionLatency(
             const std::string& name, const std::string& short_name, const std::string& group_name,
-            float prob, COMPARISON_TYPE prob_check_type, EvaluationManager& eval_man, float max_abs_value);
+            double prob, COMPARISON_TYPE prob_check_type, EvaluationManager& eval_man, float max_abs_value);
 
     float maxAbsValue() const;
