+++ src/eval/requirement/position/radarrange.cpp
@@ -37,7 +37,7 @@ namespace EvaluationRequirement
 
 PositionRadarRange::PositionRadarRange(
         const std::string& name, const std::string& short_name, const std::string& group_name,
-        EvaluationManager& eval_man, float threshold_value)
+        EvaluationManager& eval_man, double threshold_value)
     : Base(name, short_name, group_name, threshold_value, COMPARISON_TYPE::LESS_THAN_OR_EQUAL, eval_man)
 {
 }
