+++ src/eval/requirement/speed/accelerationcorrect.h
@@ -28,7 +28,7 @@ class AccelerationCorrect : public GenericDouble
 {
 public:
     AccelerationCorrect(const std::string& name, const std::string& short_name, const std::string& group_name,
-                        float prob, COMPARISON_TYPE prob_check_type, double threshold, EvaluationManager& eval_man);
+                        double prob, COMPARISON_TYPE prob_check_type, double threshold, EvaluationManager& eval_man);
 
     std::string probabilityName() const override final { return "Probability of Correct ACC"; }
     std::string probabilityNameShort() const override final { return "PCAcc"; }
