+++ src/eval/requirement/status/coastingcorrect.cpp
@@ -5,7 +5,7 @@ namespace EvaluationRequirement
 
 
 CoastingCorrect::CoastingCorrect(const std::string& name, const std::string& short_name, const std::string& group_name,
-                                     float prob, COMPARISON_TYPE prob_check_type, EvaluationManager& eval_man)
+                                     double prob, COMPARISON_TYPE prob_check_type, EvaluationManager& eval_man)
 : GenericInteger(name, short_name, group_name, prob, prob_check_type, eval_man)
 {
     result_type_ = "CoastingCorrect";
